// src/enums/codeContent.ts
export enum EnumCategories {
  CODE = 'code',
  TXT = 'text',
  MARKDOWN = 'richText',
}

export enum EnumHighlightLang {
  C = 'C/C++',
  JAVA = 'JAVA',
  JAVASCRIPT = 'JavaScript',
  PYTHON = 'Python',
  CSS = 'CSS',
  HTML = 'HTML',
  OTHER = 'other',
}

export enum EnumExpirations {
  NEVER = 'never',
  M10 = '10-M',
  H1 = '1-H',
  D1 = '1-D',
  W1 = '1-W',
  MON1 = '1-Mon',
}